+++ service/activity_service.go
@@ -7,7 +7,7 @@ import (
 )
 
 type ActivityService interface {
-	DeleteActivity(ctx context.Context, actId uint8, UID string) error
+	DeleteActivity(ctx context.Context, actId uint32, UID string) error
 	GetActivityList(ctx context.Context, UID string) ([]*web.Activity, error)
 	GetActivityCategoryById(ctx context.Context, categoryId uint) (*model.ActivityCategory, error)
 	GetActivityCategory(ctx context.Context) (*web.ResponseActivityType, error)
